Remove useless group and no_group attributes (see details)
1) These two attributes define the value to write in /etc/default/useradd
for the GROUPS key. The value indeed has been historically written by
AutoYaST, but turns out that key has been completely ignored by useradd
since SLE-12, in which he package "pwdutils" was dropped in favor of
"shadow" (see more at bsc#1099153). The latter does not honor the
configuration for GROUPS and does not offer any alternative to it.

2) In theory, those attributes were also honored by AutoYaST while
creating users during the installation process. But manual tests have
proven that was not the case in SLE-11, SLE-12 or SLE-15 (or any of
its corresponding service packs).